fix!: NumericDate parsing conformance

+++ parser_test.go
@@ -139,7 +139,7 @@ var jwtTestData = []struct {
 		"",
 		defaultKeyFunc,
 		&jwt.StandardClaims{
-			ExpiresAt: time.Now().Add(time.Second * 10).Unix(),
+			ExpiresAt: float64(time.Now().Add(time.Second * 10).Unix()),
 		},
 		true,
 		0,
